+++ src/netprotocols/__init__.py
@@ -25,6 +25,7 @@ from netprotocols.layer3.ipv6_ext import (
 )
 from netprotocols.layer4.tcp import TCP
 from netprotocols.layer4.udp import UDP
+from netprotocols.layer7.dns import DNS
 from netprotocols.packet import Packet
 from netprotocols.utils.exceptions import (
     InvalidFieldError,
@@ -41,6 +42,7 @@ __version__ = "1.1.0"
 
 __all__ = [
     "ARP",
+    "DNS",
     "TCP",
     "UDP",
     "ARPHardwareType",