+++ src/netprotocols/layer7/__init__.py
@@ -1,0 +1,3 @@
+from netprotocols.layer7.dns import DNS
+
+__all__ = ["DNS"]